fix: Chat Pane Regressions
Clicking on the toggle chat pane no longer marks all messages as read. Refactored some other code

PR-1052
