+++ src/components/useHelperApp.js
@@ -45,6 +45,9 @@ const useHelperApp = () => {
     handleToggleHelper('chat');
   };
 
+  const isOpen = (helper) => {
+    return currentHelper === helper;
+  };
 
   const HelperComponent = useMemo(() => ((props) => {
     if (!query?.helper) return null;
@@ -65,21 +68,26 @@ const useHelperApp = () => {
     );
   }), [handleToggleHelper, requestId, query.helper]);
 
-  const TagButton = ({ request }) => (
+  const TagButton = ({ request, onClick }) => (
     <FormattedMessage id="ui-rs.view.showTags">
       {ariaLabel => (
         <IconButton
           icon="tag"
           id="clickable-show-tags"
           badgeCount={request?.tags?.length ?? 0}
-          onClick={() => handleToggleTags()}
+          onClick={
+            () => {
+              handleToggleTags();
+              onClick({ open: isOpen('tags') });
+            }
+          }
           ariaLabel={ariaLabel}
         />
       )}
     </FormattedMessage>
   );
 
-  const ChatButton = ({ request }) => {
+  const ChatButton = ({ request, onClick }) => {
     const unseenNotifications = request?.notifications?.filter(notification => notification.seen === false && notification.isSender === false)?.length ?? 0;
     return (
       <FormattedMessage id="ui-rs.view.showChat">
@@ -88,7 +96,12 @@ const useHelperApp = () => {
             icon="comment"
             id="clickable-show-chat"
             badgeCount={unseenNotifications}
-            onClick={() => handleToggleChat()}
+            onClick={
+              () => {
+                handleToggleChat();
+                onClick({ open: isOpen('chat') });
+              }
+            }
             ariaLabel={ariaLabel}
           />
         )}
